+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -22,6 +22,8 @@ CHANGELOG
  * Make the `framework.uid.default_uuid_version` config option default to `7`
  * Make the `framework.uid.time_based_uuid_version` config option default to `7`
  * Make the `framework.validation.email_validation_mode` config option default to `html5`
+ * Remove the `framework.validation.enable_annotations` config option, use `framework.validation.enable_attributes` instead
+ * Remove the `framework.serializer.enable_annotations` config option, use `framework.serializer.enable_attributes` instead
 
 6.4
 ---